requirements.txt
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.1
deprecation==2.1.0
dropbox==11.33.0
google-api-core==2.8.2
google-api-python-client==2.58.0
google-auth==2.11.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.56.4
grpcio==1.48.1
httplib2==0.20.4
idna==3.3
modzy-sdk==0.8.0
packaging==21.3
ply==3.11
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-dotenv==0.20.0
requests==2.28.1
rsa==4.9
six==1.16.0
stone==3.3.1
uritemplate==4.1.1
urllib3==1.26.12
